reorganize directories for project wizard components
