+++ src/vs/workbench/browser/positronNewProjectWizard/components/steps/projectTypeStep.css
