+++ src/vs/workbench/browser/positronNewProjectWizard/components/wizardStep.css
