+++ src/vs/workbench/browser/positronNewProjectWizard/components/wizardSubStep.css
