+++ src/vs/workbench/browser/positronNewProjectWizard/interfaces/newProjectWizardStepLookup.ts
@@ -3,9 +3,9 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { NewProjectWizardStep } from 'vs/workbench/browser/positronNewProjectWizard/interfaces/newProjectWizardStep';
-import { ProjectNameLocationStep } from 'vs/workbench/browser/positronNewProjectWizard/steps/projectNameLocationStep';
-import { PythonEnvironmentStep } from 'vs/workbench/browser/positronNewProjectWizard/steps/pythonEnvironmentStep';
-import { ProjectTypeStep } from 'vs/workbench/browser/positronNewProjectWizard/steps/projectTypeStep';
+import { ProjectNameLocationStep } from 'vs/workbench/browser/positronNewProjectWizard/components/steps/projectNameLocationStep';
+import { PythonEnvironmentStep } from 'vs/workbench/browser/positronNewProjectWizard/components/steps/pythonEnvironmentStep';
+import { ProjectTypeStep } from 'vs/workbench/browser/positronNewProjectWizard/components/steps/projectTypeStep';
 
 /**
  * The NewProjectWizardStepLookup object is like a map of NewProjectWizardStep to the
